Extract selected check functions to other modules
 check: extract print function to log;
 check: extract load/store function to store;
 check: extract constants and structs to utils/structs.
