+++ CHANGELOG.rst
@@ -26,6 +26,7 @@ HEAD
   - make perun check precollect phase silent (until we figure out the better way?)
   - add streaming to the history (so it is not output when everything is done)
   - make two versions of run_jobs (one with history and one without)
+  - refactor some modules to remove unnecessary dependencies
 
 0.14.1 (2018-04-19)
 -------------------
