+++ perun/check/factory.py
@@ -1,12 +1,9 @@
-import collections
 import contextlib
-import itertools
 import os
-import operator
-import re
 
 import distutils.util as dutils
-from enum import Enum
+
+from perun.utils.structs import PerformanceChange
 
 import perun.utils.exceptions as exceptions
 import perun.profile.factory as profiles
@@ -23,135 +20,6 @@ from perun.logic.pcs import PCS
 __author__ = 'Tomas Fiedor'
 
 
-PerformanceChange = Enum(
-    'PerformanceChange', 'Degradation MaybeDegradation ' +
-                         'Unknown NoChange MaybeOptimization Optimization'
-)
-
-CHANGE_STRINGS = {
-    PerformanceChange.Degradation: 'Degradation',
-    PerformanceChange.MaybeDegradation: 'Maybe Degradation',
-    PerformanceChange.NoChange: 'No Change',
-    PerformanceChange.Unknown: 'Unknown',
-    PerformanceChange.MaybeOptimization: 'Maybe Optimization',
-    PerformanceChange.Optimization: 'Optimization'
-}
-CHANGE_COLOURS = {
-    PerformanceChange.Degradation: 'red',
-    PerformanceChange.MaybeDegradation: 'yellow',
-    PerformanceChange.NoChange: 'white',
-    PerformanceChange.Unknown: 'grey',
-    PerformanceChange.MaybeOptimization: 'cyan',
-    PerformanceChange.Optimization: 'green'
-}
-DEGRADATION_ICON = '-'
-OPTIMIZATION_ICON = '+'
-LINE_PARSING_REGEX = re.compile(
-    r"(?P<location>.+)\s"
-    r"PerformanceChange[.](?P<result>[A-Za-z]+)\s"
-    r"(?P<type>\S+)\s"
-    r"(?P<from>\S+)\s"
-    r"(?P<to>\S+)\s"
-    r"(?P<ctype>\S+)\s"
-    r"(?P<crate>\S+)\s"
-    r"(?P<minor>\S+)\s"
-    r"(?P<cmdstr>.+)"
-)
-
-
-def save_degradation_list_for(base_dir, minor_version, degradation_list):
-    """Saves the given degradation list to a minor version storage
-
-    This converts the list of degradation records to a storage-able format. Moreover,
-    this loads all of the already stored degradations. For each tuple of the change
-    location and change type, this saves only one change record.
-
-    :param str base_dir: base directory, where the degradations will be stored
-    :param str minor_version: minor version for which we are storing the degradations
-    :param degradation_list:
-    :return:
-    """
-    already_saved_changes = load_degradation_list_for(base_dir, minor_version)
-
-    list_of_registered_changes = dict()
-    already_saved_changes.extend(degradation_list)
-    for deg_info, cmdstr, source in already_saved_changes:
-        info_string = " ".join([
-            deg_info.to_storage_record(),
-            source,
-            cmdstr
-        ])
-        uid = (deg_info.location, deg_info.type)
-        list_of_registered_changes[uid] = info_string
-
-    # Sort the changes
-    to_be_stored_changes = sorted(list(list_of_registered_changes.values()))
-
-    # Store the changes in the file
-    minor_dir, minor_storage_file = store.split_object_name(base_dir, minor_version, ".changes")
-    store.touch_dir(minor_dir)
-    store.touch_file(minor_storage_file)
-    with open(minor_storage_file, 'w') as write_handle:
-        write_handle.write("\n".join(to_be_stored_changes))
-
-
-def parse_changelog_line(line):
-    """Parses one changelog record into the triple of degradation info, command string and minor.
-
-    :param str line: input line from one change log
-    :return: triple (degradation info, command string, minor version)
-    """
-    tokens = LINE_PARSING_REGEX.match(line)
-    deg_info = DegradationInfo(
-        PerformanceChange[tokens.group('result')],
-        tokens.group('type'),
-        tokens.group('location'),
-        tokens.group('from'),
-        tokens.group('to'),
-        tokens.group('ctype'),
-        float(tokens.group('crate'))
-    )
-    return deg_info, tokens.group('cmdstr'), tokens.group('minor')
-
-
-def load_degradation_list_for(base_dir, minor_version):
-    """Loads a list of degradations stored for the minor version.
-
-    This opens a file in the .perun/objects directory in the minor version subdirectory with the
-    extension ".changes". The file is basically a log of degradation records separated by
-    white spaces in ascii coding.
-
-    :param str base_dir: directory to the storage of the objects
-    :param str minor_version:
-    :return: list of triples (DegradationInfo, command string, minor version source)
-    """
-    minor_dir, minor_storage_file = store.split_object_name(base_dir, minor_version, ".changes")
-    store.touch_dir(minor_dir)
-    store.touch_file(minor_storage_file)
-    with open(minor_storage_file, 'r') as read_handle:
-        lines = read_handle.readlines()
-
-    degradation_list = []
-    for line in lines:
-        parsed_triple = parse_changelog_line(line.strip())
-        degradation_list.append(parsed_triple)
-    return degradation_list
-
-
-def count_degradations_per_group(degradation_list):
-    """Counts the number of optimizations and degradations
-
-    :param list degradation_list: list of tuples of (degradation info, cmdstr, minor version)
-    :return: dictionary mapping change strings to its counts
-    """
-    # Get only degradation results
-    changes = map(operator.attrgetter('result'), map(operator.itemgetter(0), degradation_list))
-    # Transform the enum into a string
-    changes = list(map(operator.attrgetter('name'), changes))
-    counts = dict(collections.Counter(changes))
-    return counts
-
-
 def profiles_to_queue(pcs, minor_version):
     """Retrieves the list of profiles corresponding to minor version and transforms them to map.
 
@@ -165,127 +33,6 @@ def profiles_to_queue(pcs, minor_version):
     return {
         profile.config_tuple: profile for profile in minor_version_profiles
     }
-
-
-def get_degradation_change_colours(degradation_result):
-    """Returns the tuple of two colours w.r.t degradation results.
-
-    If the change was optimization (or possible optimization) then we print the first model as
-    red and the other by green (since we went from better to worse model). On the other hand if the
-    change was degradation, then we print the first one green (was better) and the other as red
-    (is now worse). Otherwise (for Unknown and no change) we keep the stuff yellow, though this
-    is not used at all
-
-    :param PerformanceChange degradation_result: change of the performance
-    :returns: tuple of (from model string colour, to model string colour)
-    """
-    if degradation_result in (
-            PerformanceChange.Optimization, PerformanceChange.MaybeOptimization
-    ):
-        return 'red', 'green'
-    elif degradation_result in (
-            PerformanceChange.Degradation, PerformanceChange.MaybeDegradation
-    ):
-        return 'green', 'red'
-    else:
-        return 'yellow', 'yellow'
-
-
-def print_short_summary_of_degradations(degradation_list):
-    """Prints a short string representing the summary of the found changes.
-
-    This prints a short statistic of found degradations and short summary string.
-
-    :param list degradation_list:
-        list of tuples (degradation info, command string, source minor version)
-    """
-    counts = count_degradations_per_group(degradation_list)
-
-    print("")
-    print_short_change_string(counts)
-    print(" {} optimizations({}), {} degradations({})".format(
-        counts.get('Optimization', 0), OPTIMIZATION_ICON,
-        counts.get('Degradation', 0), DEGRADATION_ICON
-    ))
-
-
-def print_short_change_string(counts):
-    """Prints short string representing a summary of the given degradation list.
-
-    This prints a short string of form representing a summary of found optimizations (+) and
-    degradations (-) in the given degradation list. Uncertain optimizations and degradations
-    are omitted. The string can e.g. look as follows:
-
-    ++++-----
-
-    :param dict counts: dictionary mapping found string changes into their counts
-    """
-    overall_changes = sum(counts.values())
-    print("{} change{}".format(
-        overall_changes, "s" if overall_changes != 1 else ""
-    ), end='')
-    if overall_changes > 0:
-        print(" | ", end='')
-        log.cprint(
-            OPTIMIZATION_ICON*counts.get('Optimization', 0),
-            CHANGE_COLOURS[PerformanceChange.Optimization],
-            attrs=['bold']
-        )
-        log.cprint(
-            DEGRADATION_ICON*counts.get('Degradation', 0),
-            CHANGE_COLOURS[PerformanceChange.Degradation],
-            attrs=['bold']
-        )
-    print("")
-
-
-def print_list_of_degradations(degradation_list):
-    """Prints list of found degradations grouped by location
-
-    Currently this is hardcoded and prints the list of degradations as follows:
-
-    at {loc}:
-      {result} from {from} -> to {to}
-
-    :param list degradation_list: list of found degradations
-    """
-    def keygetter(item):
-        """Returns the location of the degradation from the tuple
-
-        :param tuple item: tuple of (degradation result, cmd string, source minor version)
-        :return: location of the degradation used for grouping
-        """
-        return item[0].location
-
-    # Group by location
-    degradation_list.sort(key=keygetter)
-    for location, changes in itertools.groupby(degradation_list, keygetter):
-        # Print the location
-        print("at", end='')
-        log.cprint(' {}'.format(location), 'white', attrs=['bold'])
-        print(":")
-        # Iterate and print all of the infos
-        for deg_info, _, __ in changes:
-            log.cprint(
-                '  {}'.format(CHANGE_STRINGS[deg_info.result]),
-                CHANGE_COLOURS[deg_info.result], attrs=['bold']
-            )
-            if deg_info.result != PerformanceChange.NoChange:
-                from_colour, to_colour = get_degradation_change_colours(deg_info.result)
-                print(' from: ', end='')
-                log.cprint('{}'.format(deg_info.from_baseline), from_colour, attrs=[])
-                print(' -> to: ', end='')
-                log.cprint('{}'.format(deg_info.to_target), to_colour, attrs=[])
-                if deg_info.confidence_type != 'no':
-                    print(' (with confidence ', end='')
-                    log.cprint(
-                        '{} = {}'.format(
-                            deg_info.confidence_type, deg_info.confidence_rate),
-                        'white', attrs=['bold']
-                    )
-                    print(')', end='')
-            print('')
-    print("")
 
 
 @decorators.static_variables(minor_version_cache=set())
@@ -355,9 +102,9 @@ def degradation_in_minor(minor_version, quiet=False):
                 del target_profile_queue[target_profile.config_tuple]
 
         # Store the detected degradation
-        save_degradation_list_for(pcs.get_object_directory(), minor_version, detected_changes)
+        store.save_degradation_list_for(pcs.get_object_directory(), minor_version, detected_changes)
     if not quiet:
-        print_list_of_degradations(detected_changes)
+        log.print_list_of_degradations(detected_changes)
     return detected_changes
 
 
@@ -374,12 +121,12 @@ def degradation_in_history(head):
         for minor_version in vcs.walk_minor_versions(pcs.vcs_type, pcs.vcs_path, head):
             history.progress_to_next_minor_version(minor_version)
             newly_detected_changes = degradation_in_minor(minor_version.checksum, True)
-            print_short_change_string(count_degradations_per_group(newly_detected_changes))
+            log.print_short_change_string(log.count_degradations_per_group(newly_detected_changes))
             history.finish_minor_version(minor_version, newly_detected_changes)
-            print_list_of_degradations(newly_detected_changes)
+            log.print_list_of_degradations(newly_detected_changes)
             detected_changes.extend(newly_detected_changes)
             history.flush(with_border=True)
-    print_short_summary_of_degradations(detected_changes)
+    log.print_short_summary_of_degradations(detected_changes)
     return detected_changes
 
 
@@ -429,8 +176,8 @@ def degradation_between_files(baseline_file, target_file, minor_version):
 
     # Store the detected changes for given minor version
     pcs = PCS(store.locate_perun_dir_on(os.getcwd()))
-    save_degradation_list_for(pcs.get_object_directory(), target_minor_version, detected_changes)
-    print_short_summary_of_degradations(detected_changes)
+    store.save_degradation_list_for(pcs.get_object_directory(), target_minor_version, detected_changes)
+    log.print_short_summary_of_degradations(detected_changes)
 
 
 def is_rule_applicable_for(rule, configuration):
@@ -514,60 +261,3 @@ def get_strategies_for_configuration(profile):
             yield method
 
 
-class DegradationInfo(object):
-    """The returned results for performance check methods
-
-    :ivar PerformanceChange result: result of the performance change, either can be optimization,
-        degradation, no change, or certain type of unknown
-    :ivar str type: string representing the type of the degradation, e.g. "order" degradation
-    :ivar str location: location, where the degradation has happened
-    :ivar str from_baseline: value or model representing the baseline, i.e. from which the new
-        version was optimized or degraded
-    :ivar str to_target: value or model representing the target, i.e. to which the new version was
-        optimized or degraded
-    :ivar str confidence_type: type of the confidence we have in the detected degradation, e.g. r^2
-    :ivar float confidence_rate: value of the confidence we have in the detected degradation
-    """
-
-    def __init__(self, res, t, loc, fb, tt, ct="no", cr=0.0):
-        """Each degradation consists of its results, the location, where the change has happened
-        (this is e.g. the unique id of the resource, like function or concrete line), then the pair
-        of best models for baseline and target, and the information about confidence.
-
-        E.g. for models we can use coefficient of determination as some kind of confidence, e.g. the
-        higher the confidence the more likely we predicted successfully the degradation or
-        optimization.
-
-        :param PerformanceChange res: result of the performance change, either can be optimization,
-            degradation, no change, or certain type of unknown
-        :param str t: string representing the type of the degradation, e.g. "order" degradation
-        :param str loc: location, where the degradation has happened
-        :param str fb: value or model representing the baseline, i.e. from which the new version was
-            optimized or degraded
-        :param str tt: value or model representing the target, i.e. to which the new version was
-            optimized or degraded
-        :param str ct: type of the confidence we have in the detected degradation, e.g. r^2
-        :param float cr: value of the confidence we have in the detected degradation
-        """
-        self.result = res
-        self.type = t
-        self.location = loc
-        self.from_baseline = fb
-        self.to_target = tt
-        self.confidence_type = ct
-        self.confidence_rate = cr
-
-    def to_storage_record(self):
-        """Transforms the degradation info to a storage_record
-
-        :return: string representation of the degradation as a stored record in the file
-        """
-        return "{} {} {} {} {} {} {}".format(
-            self.location,
-            self.result,
-            self.type,
-            self.from_baseline,
-            self.to_target,
-            self.confidence_type,
-            self.confidence_rate
-        )
